CLI-562 Support remove/undo in integrations framework
