+++ tests/unit/cli/commands/integrate/_common/registry.test.ts
@@ -455,470 +455,6 @@ describe('declarative integration framework', () => {
     await resource.apply(makeContext(state, tempDir));
     expect(await readFile(targetPath, 'utf-8')).toBe('#!/bin/sh\n# sonar-managed\necho sonar\n');
   });
-
-  it('updates text snippets in existing files and escapes marker characters', async () => {
-    const state = getDefaultState('test');
-    const context = makeContext(state, tempDir);
-    const targetPath = join(tempDir, 'existing.txt');
-    await writeFile(
-      targetPath,
-      [
-        'before',
-        '# sonar:begin [feature]',
-        'old content',
-        '# sonar:end [feature]',
-        'after',
-        '',
-      ].join('\n'),
-    );
-    const resource = textSnippet({
-      id: 'feature',
-      targetPath,
-      content: 'new content',
-      startMarker: '# sonar:begin [feature]',
-      endMarker: '# sonar:end [feature]',
-    });
-
-    await resource.apply(context);
-
-    expect(await readFile(targetPath, 'utf-8')).toBe(
-      [
-        'before',
-        '# sonar:begin [feature]',
-        'new content',
-        '# sonar:end [feature]',
-        'after',
-        '',
-      ].join('\n'),
-    );
-    expect(await resource.isApplied(context)).toBe(true);
-  });
-
-  it('appends text snippets to non-empty files and reports missing snippets as not applied', async () => {
-    const state = getDefaultState('test');
-    const context = makeContext(state, tempDir);
-    const targetPath = join(tempDir, 'append.txt');
-    const resource = textSnippet({
-      id: 'append',
-      targetPath,
-      content: 'managed content',
-      startMarker: '# sonar:begin append',
-    });
-
-    expect(await resource.isApplied(context)).toBe(false);
-
-    await writeFile(targetPath, 'existing content\n');
-    await resource.apply(context);
-
-    expect(await readFile(targetPath, 'utf-8')).toBe(
-      [
-        'existing content',
-        '',
-        '# sonar:begin append',
-        'managed content',
-        '# sonar:end append',
-        '',
-      ].join('\n'),
-    );
-  });
-
-  it('replaces legacy text snippets that only contain the start marker', async () => {
-    const state = getDefaultState('test');
-    const context = makeContext(state, tempDir);
-    const targetPath = join(tempDir, 'legacy.txt');
-    await writeFile(
-      targetPath,
-      [
-        '#!/bin/sh',
-        '# sonar:begin append',
-        'old managed content',
-        '"$SONAR_BIN" hook git-pre-commit',
-        '',
-      ].join('\n'),
-    );
-    const resource = textSnippet({
-      id: 'append',
-      targetPath,
-      content: 'managed content',
-      startMarker: '# sonar:begin append',
-    });
-
-    await resource.apply(context);
-
-    expect(await readFile(targetPath, 'utf-8')).toBe(
-      ['#!/bin/sh', '# sonar:begin append', 'managed content', '# sonar:end append', ''].join('\n'),
-    );
-  });
-
-  it('skips operations when shouldApply returns false', async () => {
-    const state = getDefaultState('test');
-    let called = false;
-    const feature: FeatureDeclaration = {
-      id: 'feature',
-      displayName: 'Feature',
-      operations: [
-        {
-          id: 'operation',
-          shouldApply: () => false,
-          apply: () => {
-            called = true;
-          },
-        },
-      ],
-    };
-    const integration = makeIntegration({ features: [feature] });
-
-    const installed = await applyAndRecord(
-      installer,
-      makeContext(state, tempDir),
-      integration,
-      feature,
-    );
-
-    expect(called).toBe(false);
-    expect(installed.operations).toEqual([]);
-  });
-
-  it('passes update execution mode to feature operations', async () => {
-    const state = getDefaultState('test');
-    let called = false;
-    const feature: FeatureDeclaration = {
-      id: 'feature',
-      displayName: 'Feature',
-      operations: [
-        {
-          id: 'operation',
-          shouldApply: (context) => context.executionMode === 'install',
-          apply: () => {
-            called = true;
-          },
-        },
-      ],
-    };
-    const integration = makeIntegration({ features: [feature] });
-
-    const installed = await installer.applyAndRecordFeatures(
-      state,
-      integration,
-      [
-        {
-          feature,
-          targetRoot: tempDir,
-          scope: 'project',
-        },
-      ],
-      { executionMode: 'update' },
-    );
-
-    expect(called).toBe(false);
-    expect(installed[0]?.operations).toEqual([]);
-  });
-
-  it('records multiple features under one installed integration', async () => {
-    const state = getDefaultState('test');
-    const firstFeature: FeatureDeclaration = {
-      id: 'first',
-      displayName: 'First',
-      operations: [{ id: 'first-operation', apply: () => undefined }],
-    };
-    const secondFeature: FeatureDeclaration = {
-      id: 'second',
-      displayName: 'Second',
-      operations: [{ id: 'second-operation', apply: () => undefined }],
-    };
-    const integration = makeIntegration({ features: [firstFeature, secondFeature] });
-    const context = makeContext(state, tempDir);
-
-    await applyAndRecord(installer, context, integration, firstFeature);
-    await applyAndRecord(installer, context, integration, secondFeature);
-
-    expect(state.integrations.installed).toHaveLength(1);
-    expect(state.integrations.installed[0].integrationId).toBe('test-integration');
-    expect(state.integrations.installed[0].features.map((feature) => feature.featureId)).toEqual([
-      'first',
-      'second',
-    ]);
-  });
-
-  it('records the same feature for different targets under one installed integration', async () => {
-    const state = getDefaultState('test');
-    const feature: FeatureDeclaration = {
-      id: 'feature',
-      displayName: 'Feature',
-      operations: [{ id: 'operation', apply: () => undefined }],
-    };
-    const integration = makeIntegration({ features: [feature] });
-
-    await applyAndRecord(
-      installer,
-      makeContext(state, join(tempDir, 'project')),
-      integration,
-      feature,
-    );
-    await applyAndRecord(
-      installer,
-      makeContext(state, join(tempDir, 'global')),
-      integration,
-      feature,
-    );
-
-    expect(state.integrations.installed).toHaveLength(1);
-    expect(state.integrations.installed[0].features).toHaveLength(2);
-    expect(
-      state.integrations.installed[0].features.map((entry) => entry.targetRoot).sort(),
-    ).toEqual([join(tempDir, 'global'), join(tempDir, 'project')]);
-  });
-
-  it('prunes stale feature state when declarations change from operations to resources', async () => {
-    const state = getDefaultState('test');
-    const context = makeContext(state, tempDir);
-    const legacyFeature: FeatureDeclaration = {
-      id: 'feature',
-      displayName: 'Feature',
-      operations: [{ id: 'legacy-operation', apply: () => undefined }],
-    };
-    const currentFeature: FeatureDeclaration = {
-      id: 'feature',
-      displayName: 'Feature',
-      resources: [
-        jsonPatch({
-          id: 'json',
-          targetPath: join(tempDir, 'settings.json'),
-          patch: () => ({ enabled: true }),
-        }),
-      ],
-    };
-
-    await applyAndRecord(
-      installer,
-      context,
-      makeIntegration({ features: [legacyFeature] }),
-      legacyFeature,
-    );
-    const installed = await applyAndRecord(
-      installer,
-      context,
-      makeIntegration({ features: [currentFeature] }),
-      currentFeature,
-    );
-
-    expect(installed.operations).toEqual([]);
-    expect(installed.resources).toMatchObject([
-      {
-        id: 'json',
-        resourceType: 'json-patch',
-        path: join(tempDir, 'settings.json'),
-      },
-    ]);
-  });
-
-  it('prunes stale shared dependency state when no installed feature references it', async () => {
-    const state = getDefaultState('test');
-    const context = makeContext(state, tempDir);
-    const legacyFeature: FeatureDeclaration = {
-      id: 'feature',
-      displayName: 'Feature',
-      dependencies: [
-        sonarSourceBinary({
-          id: 'binary',
-          binary: SonarSourceBinary.SonarSecrets,
-        }),
-      ],
-    };
-    const currentFeature: FeatureDeclaration = {
-      id: 'feature',
-      displayName: 'Feature',
-      resources: [
-        wholeFile({
-          id: 'current',
-          targetPath: join(tempDir, 'current.txt'),
-          content: 'current',
-        }),
-      ],
-    };
-
-    await applyAndRecord(
-      installer,
-      context,
-      makeIntegration({ features: [legacyFeature] }),
-      legacyFeature,
-    );
-    const installed = await applyAndRecord(
-      installer,
-      context,
-      makeIntegration({ features: [currentFeature] }),
-      currentFeature,
-    );
-
-    expect(installed.dependencies).toEqual([]);
-    expect(state.dependencies.installed).toEqual([]);
-  });
-
-  it('checks SonarSource binary dependencies by their descriptor', async () => {
-    const state = getDefaultState('test');
-    const binaryPath = join(tempDir, 'bin', 'sonar-secrets');
-    const dependency = sonarSourceBinary({
-      id: 'binary',
-      binary: SonarSourceBinary.SonarSecrets,
-    });
-    const context = makeContext(state, tempDir);
-
-    expect(await dependency.isInstalled(context)).toBe(false);
-
-    resolveBinaryPathSpy.mockReturnValue(binaryPath);
-
-    expect(await dependency.isInstalled(context)).toBe(true);
-
-    const applied = await dependency.installOrUpdate(context);
-
-    expect(installBinarySpy).toHaveBeenCalledWith(SonarSourceBinary.SonarSecrets.spec);
-    expect(applied).toEqual({
-      id: 'binary',
-      dependencyType: 'sonarsource-binary',
-      version: SonarSourceBinary.SonarSecrets.spec.version,
-      path: binaryPath,
-    });
-  });
-
-  it('reports whether dependencies, resources, and operations need to be applied', async () => {
-    const state = getDefaultState('test');
-    const feature: FeatureDeclaration = {
-      id: 'feature',
-      displayName: 'Feature',
-      dependencies: [
-        sonarSourceBinary({
-          id: 'dependency',
-          version: '1',
-          binary: SonarSourceBinary.SonarSecrets,
-        }),
-      ],
-      resources: [
-        wholeFile({
-          id: 'resource',
-          version: '1',
-          targetPath: join(tempDir, 'file.txt'),
-          content: 'content',
-        }),
-      ],
-      operations: [{ id: 'operation', version: '1', apply: () => undefined }],
-    };
-    const integration = makeIntegration({ features: [feature] });
-    const context = makeContext(state, tempDir);
-
-    expect(await installer.dependencyNeedsInstall(context, feature.dependencies![0])).toBe(true);
-    expect(await installer.resourceNeedsApply(context, undefined, feature.resources![0])).toBe(
-      true,
-    );
-    expect(installer.operationNeedsApply(undefined, feature.operations![0])).toBe(true);
-    expect(
-      installer.operationNeedsApply(undefined, { id: 'unversioned', apply: () => undefined }),
-    ).toBe(true);
-
-    const installed = await applyAndRecord(installer, context, integration, feature);
-    const found = installer.findInstalledFeature(state, context, integration, feature);
-    resolveBinaryPathSpy.mockReturnValue(join(tempDir, 'bin', 'sonar-secrets'));
-
-    expect(found?.featureId).toBe(installed.featureId);
-    expect(await installer.dependencyNeedsInstall(context, feature.dependencies![0])).toBe(false);
-    expect(await installer.resourceNeedsApply(context, installed, feature.resources![0])).toBe(
-      false,
-    );
-    expect(installer.operationNeedsApply(installed, feature.operations![0])).toBe(false);
-    expect(
-      await installer.dependencyNeedsInstall(
-        context,
-        sonarSourceBinary({
-          id: 'dependency',
-          version: '2',
-          binary: SonarSourceBinary.SonarSecrets,
-        }),
-      ),
-    ).toBe(true);
-    expect(
-      await installer.resourceNeedsApply(
-        context,
-        installed,
-        wholeFile({
-          id: 'resource',
-          version: '2',
-          targetPath: join(tempDir, 'file.txt'),
-          content: 'content',
-        }),
-      ),
-    ).toBe(true);
-    expect(
-      await installer.resourceNeedsApply(
-        context,
-        installed,
-        wholeFile({
-          id: 'resource',
-          version: '1',
-          targetPath: join(tempDir, 'file.txt'),
-          content: 'updated content',
-        }),
-      ),
-    ).toBe(true);
-    expect(
-      installer.operationNeedsApply(installed, { ...feature.operations![0], version: '2' }),
-    ).toBe(true);
-  });
-
-  it('fails when JSON files contain invalid content', async () => {
-    const state = getDefaultState('test');
-    const context = makeContext(state, tempDir);
-    const jsonPath = join(tempDir, 'settings.json');
-    await writeFile(jsonPath, '{ invalid json');
-    const jsonResource = jsonPatch({
-      id: 'json-invalid',
-      targetPath: jsonPath,
-      defaultValue: { fallback: true },
-      patch: (document) => ({ ...(document as Record<string, unknown>), enabled: true }),
-    });
-
-    expect(jsonResource.apply(context)).rejects.toThrow(
-      `${jsonPath} contains invalid JSON. Please fix or delete it and re-run.`,
-    );
-    expect(jsonResource.isApplied(context)).rejects.toThrow(
-      `${jsonPath} contains invalid JSON. Please fix or delete it and re-run.`,
-    );
-  });
-
-  it('fails when TOML files contain invalid content', async () => {
-    const state = getDefaultState('test');
-    const context = makeContext(state, tempDir);
-    const tomlPath = join(tempDir, 'config.toml');
-    await writeFile(tomlPath, '= not valid toml =');
-    const tomlResource = tomlPatch({
-      id: 'toml-invalid',
-      targetPath: tomlPath,
-      defaultValue: {},
-      patch: (document) => document,
-    });
-
-    expect(tomlResource.apply(context)).rejects.toThrow(
-      `${tomlPath} contains invalid TOML. Please fix or delete it and re-run.`,
-    );
-    expect(tomlResource.isApplied(context)).rejects.toThrow(
-      `${tomlPath} contains invalid TOML. Please fix or delete it and re-run.`,
-    );
-  });
-
-  it('uses defaults when YAML files contain invalid content', async () => {
-    const state = getDefaultState('test');
-    const context = makeContext(state, tempDir);
-    const yamlPath = join(tempDir, 'settings.yml');
-    await writeFile(yamlPath, 'invalid: [yaml');
-    const yamlResource = yamlPatch({
-      id: 'yaml-invalid',
-      targetPath: yamlPath,
-      patch: (document) => ({ ...(document as Record<string, unknown>), enabled: true }),
-    });
-
-    await yamlResource.apply(context);
-
-    expect(await readFile(yamlPath, 'utf-8')).toBe('enabled: true\n');
-    expect(await yamlResource.isApplied(context)).toBe(true);
-  });
 });
 
 function makeIntegration<TOptions = Record<string, unknown>>(
